Your-Stance/requirements.txt
django-allauth==0.24.1
Django==1.9.6
requests==2.9.1
requests-oauthlib==0.6.1
dj-database-url==0.4.0
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
django-debug-toolbar==1.4
django-appconf==1.0.1
django-imagekit==3.3
django-storages==1.4
pilkit==1.1.13
Pillow==3.1.1
django-model-utils==2.4
boto==2.39.0
django-ses==0.7.1
django-loginas==0.2.0
rollbar==0.13.2
django-nose==1.4.4
pycodestyle==2.0.0
pyflakes==1.2.3
mccabe==0.5.0
python-twitter==2.2
sqlparse==0.1.19
django-reversion==2.0.6